+++ packages/typespec-rust/src/tcgcadapter/adapter.ts
@@ -355,7 +355,13 @@ export class Adapter {
     if (<tcgc.UsageFlags>(model.usage & tcgc.UsageFlags.Input) === tcgc.UsageFlags.Input) {
       modelFlags |= rust.ModelFlags.Input;
     }
-    if (<tcgc.UsageFlags>(model.usage & tcgc.UsageFlags.Output) === tcgc.UsageFlags.Output) {
+
+    // include error and LRO polling types as output types
+    if (
+      <tcgc.UsageFlags>(model.usage & tcgc.UsageFlags.Output) === tcgc.UsageFlags.Output ||
+      <tcgc.UsageFlags>(model.usage & tcgc.UsageFlags.Exception) === tcgc.UsageFlags.Exception ||
+      <tcgc.UsageFlags>(model.usage & tcgc.UsageFlags.LroPolling) === tcgc.UsageFlags.LroPolling
+    ) {
       modelFlags |= rust.ModelFlags.Output;
     }