+++ packages/typespec-rust/CHANGELOG.md
@@ -11,6 +11,7 @@
 * Using `Azure.Core.eTag` as a header parameter no longer causes an internal error.
 * Fixed bad codegen for remaining cases when an operation is annotated with `Access.internal` (the original fix for this was incomplete).
 * Fixed incorrect handling of `@alternateType` decorator when replacing scalar types.
+* Don't derive `Default` on internal structs used for spread params.
 
 ### Other Changes